extract NotificationListItemBase base class
