+++ packages/fiori/src/NotifactionOverflowActionsPopover.hbs
@@ -4,7 +4,7 @@
 	horizontal-align="Right"
 	no-arrow
 >
-	<div class="ui5-nli-overflow-list">
+	<div class="ui5-notification-overflow-list">
 		{{#each overflowActions}}
 			<ui5-button
 				icon="{{this.icon}}"
