+++ packages/fiori/src/NotificationListItem.hbs
@@ -18,6 +18,7 @@
 				@click="{{_onBtnOverflowClick}}"
 				class="ui5-nli-overflow-btn"
 				title="{{overflowBtnTitle}}"
+				overflow-btn
 			></ui5-button>
 		{{else}}
 			{{#each standardActions}}
