+++ packages/fiori/src/NotificationListItem.js
@@ -3,19 +3,10 @@ import { isSpace, isEnter } from "@ui5/webcomponents-base/dist/Keys.js";
 import { getI18nBundle, fetchI18nBundle } from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import ResizeHandler from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.js";
 import { isIE } from "@ui5/webcomponents-base/dist/Device.js";
-import { getRTL } from "@ui5/webcomponents-base/dist/config/RTL.js";
 
-import ListItemBase from "@ui5/webcomponents/dist/ListItemBase.js";
 import Button from "@ui5/webcomponents/dist/Button.js";
 import Icon from "@ui5/webcomponents/dist/Icon.js";
-import Priority from "@ui5/webcomponents/dist/types/Priority.js";
-
-// Icons
-import "@ui5/webcomponents-icons/dist/icons/decline.js";
-import "@ui5/webcomponents-icons/dist/icons/message-success.js";
-import "@ui5/webcomponents-icons/dist/icons/message-error.js";
-import "@ui5/webcomponents-icons/dist/icons/message-warning.js";
-import "@ui5/webcomponents-icons/dist/icons/overflow.js";
+import NotificationListItemBase from "./NotificationListItemBase.js";
 
 // Texts
 import {
@@ -26,17 +17,9 @@ import {
 
 // Templates
 import NotificationListItemTemplate from "./generated/templates/NotificationListItemTemplate.lit.js";
-import NotificationListItemPopoverTemplate from "./generated/templates/NotificationListItemPopoverTemplate.lit.js";
 
 // Styles
 import NotificationListItemCss from "./generated/themes/NotificationListItem.css.js";
-import NotifactionListItemPopoverCss from "./generated/themes/NotifactionListItemPopover.css.js";
-
-const PRIORITY_ICONS_MAP = {
-	"High": "message-error",
-	"Medium": "message-warning",
-	"Low": "message-success",
-};
 
 const MAX_WRAP_HEIGHT = 32; // px.
 
@@ -47,36 +30,6 @@ const metadata = {
 	tag: "ui5-li-notification",
 	managedSlots: true,
 	properties: /** @lends sap.ui.webcomponents.fiori.NotificationListItem.prototype */ {
-		/**
-		 * Defines the <code>heading</code>.
-		 * @type {string}
-		 * @defaultvalue ""
-		 * @public
-		 */
-		heading: {
-			type: String,
-		},
-
-		/**
-		 * Defines the <code>priority</code> of the notification.
-		 * @type {Priority}
-		 * @defaultvalue "None"
-		 * @public
-		 */
-		priority: {
-			type: Priority,
-			defaultValue: Priority.None,
-		},
-
-		/**
-		 * Defines if a <code>close</code> button should be displayed.
-		 * @type {boolean}
-		 * @defaultvalue false
-		 * @public
-		 */
-		showClose: {
-			type: Boolean,
-		},
 
 		/**
 		 * Defines if the <code>heading</code> and <code>decription</code> should truncate,
@@ -88,7 +41,6 @@ const metadata = {
 		truncate: {
 			type: Boolean,
 		},
-
 
 		/**
 		 * Defines if the <code>notification</code> is new or has been already read.
@@ -121,19 +73,6 @@ const metadata = {
 		},
 	},
 	slots: /** @lends sap.ui.webcomponents.fiori.NotificationListItem.prototype */ {
-
-		/**
-		 * Defines the actions, displayed in the <code>ui5-li-notification</code>.
-		 * <br><br>
-		 * <b>Note:</b> Consider using the <code>ui5-notification-overflow-action</code>.
-		 *
-		 * @type {HTMLElement}
-		 * @slot
-		 * @public
-		 */
-		actions: {
-			type: HTMLElement,
-		},
 
 		/**
 		 * Defines the avatar, displayed in the <code>ui5-li-notification</code>.
@@ -178,7 +117,6 @@ const metadata = {
 		},
 	},
 	events: /** @lends sap.ui.webcomponents.fiori.NotificationListItem.prototype */ {
-		_close: {},
 		_press: {},
 	},
 };
@@ -213,13 +151,13 @@ const metadata = {
  * @constructor
  * @author SAP SE
  * @alias sap.ui.webcomponents.fiori.NotificationListItem
- * @extends ListItemBase
+ * @extends NotificationListItemBase
  * @tagname ui5-li-notification
  * @appenddocs NotificationOverflowAction
  * @since 1.0.0-rc.8
  * @public
  */
-class NotificationListItem extends ListItemBase {
+class NotificationListItem extends NotificationListItemBase {
 	constructor() {
 		super();
 
@@ -251,14 +189,6 @@ class NotificationListItem extends ListItemBase {
 		return NotificationListItemTemplate;
 	}
 
-	static get staticAreaTemplate() {
-		return NotificationListItemPopoverTemplate;
-	}
-
-	static get staticAreaStyles() {
-		return NotifactionListItemPopoverCss;
-	}
-
 	static async onDefine() {
 		await Promise.all([
 			Button.define(),
@@ -275,15 +205,6 @@ class NotificationListItem extends ListItemBase {
 		ResizeHandler.deregister(this, this.onResizeBind);
 	}
 
-	get footerItems() {
-		return this.footnotes.map((el, idx, arr) => {
-			return {
-				slotName: el._individualSlot,
-				showDivider: idx !== arr.length - 1,
-			};
-		});
-	}
-
 	get hasHeading() {
 		return !!this.heading.length;
 	}
@@ -294,14 +215,6 @@ class NotificationListItem extends ListItemBase {
 
 	get hasFootNotes() {
 		return !!this.footnotes.length;
-	}
-
-	get hasPriority() {
-		return this.priority !== Priority.None;
-	}
-
-	get priorityIcon() {
-		return PRIORITY_ICONS_MAP[this.priority];
 	}
 
 	get showMoreText() {
@@ -368,37 +281,11 @@ class NotificationListItem extends ListItemBase {
 		return description.offsetHeight < description.scrollHeight;
 	}
 
-	get overflowButtonDOM() {
-		return this.shadowRoot.querySelector(".ui5-nli-overflow-btn");
-	}
-
-	get showOverflow() {
-		return !!this.overflowActions.length;
-	}
-
-	get overflowActions() {
-		if (this.actions.length <= 1) {
-			return [];
-		}
-
-		return this.actionsInfo;
-	}
-
-	get standardActions() {
-		if (this.actions.length > 1) {
-			return [];
-		}
-
-		return this.actionsInfo;
-	}
-
-	get actionsInfo() {
-		return this.actions.map(action => {
+	get footerItems() {
+		return this.footnotes.map((el, idx, arr) => {
 			return {
-				icon: action.icon,
-				text: action.text,
-				press: this._onCustomActionPress.bind(this),
-				refItemid: action._id,
+				slotName: el._individualSlot,
+				showDivider: idx !== arr.length - 1,
 			};
 		});
 	}
@@ -420,11 +307,6 @@ class NotificationListItem extends ListItemBase {
 		return ids.join(" ");
 	}
 
-
-	get rtl() {
-		return getRTL() ? "rtl" : undefined;
-	}
-
 	get classes() {
 		return {
 			content: {
@@ -435,9 +317,7 @@ class NotificationListItem extends ListItemBase {
 
 	/**
 	 * Event handlers
-	 *
 	 */
-
 	_onclick(event) {
 		if (event.isMarked === "button") {
 			return;
@@ -449,20 +329,8 @@ class NotificationListItem extends ListItemBase {
 		this._showMorePressed = !this._showMorePressed;
 	}
 
-	_onBtnCloseClick() {
-		this.fireEvent("_close", { item: this });
-	}
-
-	_onBtnOverflowClick() {
-		this.openOverflow();
-	}
-
 	_onkeydown(event) {
 		super._onkeydown(event);
-
-		if (isSpace(event)) {
-			event.preventDefault();
-		}
 
 		if (isEnter(event)) {
 			this.fireItemPress(event);
@@ -505,38 +373,6 @@ class NotificationListItem extends ListItemBase {
 		}
 
 		this._showMore = false;
-	}
-
-	_onCustomActionPress(event) {
-		const refItemId = event.target.getAttribute("data-ui5-external-action-item-id");
-
-		if (refItemId) {
-			this.getActionByID(refItemId).fireEvent("click", {
-				targetRef: event.target,
-			}, true);
-
-			this.closeOverflow();
-		}
-	}
-
-
-	getActionByID(id) {
-		return this.actions.find(action => action._id === id);
-	}
-
-	async openOverflow() {
-		const overflowPopover = await this.getOverflowPopover();
-		overflowPopover.openBy(this.overflowButtonDOM);
-	}
-
-	async closeOverflow() {
-		const overflowPopover = await this.getOverflowPopover();
-		overflowPopover.close();
-	}
-
-	async getOverflowPopover() {
-		const staticAreaItem = await this.getStaticAreaItemDomRef();
-		return staticAreaItem.querySelector(".ui5-notification-overflow-popover");
 	}
 }
