+++ packages/fiori/src/themes/NotifactionOverflowActionsPopover.css
@@ -1,4 +1,4 @@
-.ui5-nli-overflow-list {
+.ui5-notification-overflow-list {
 	display: flex;
 	flex-direction: column;
 	padding: 0 0.5rem;
